requirements.txt
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
bcrypt==3.2.0
boto3==1.16.47
botocore==1.19.48
cffi==1.14.4
click==7.1.2
constantly==15.1.0
cryptography==3.2.1
Django==3.1.5
flask==1.1.2
flask-cors==3.0.9
flask-sqlalchemy==2.4.4
hyperlink==20.0.1
idna==3.1
image==1.5.33
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
jwt==1.1.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
numpy==1.19.4
observable==1.0.3
pandas==1.2.0
Pillow==8.1.0
protobuf==3.14.0
PyHamcrest==2.0.2
python-dateutil==2.8.1
pytz==2020.5
s3transfer==0.3.3
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.26.2
Werkzeug==1.0.1
zope.interface==5.2.0
